questions/min_tokens.py
"""
Problem: Minimum Tokens for Total

Description:
Given a set of n distinct tokens, each assigned a positive integer value, the goal is to find the minimum number of tokens needed to reach a total value of x. If it's impossible to achieve the total using the available tokens, the function should return a value indicating this impossibility.

Function Signature:
def min_tokens_for_total(n: int, x: int, token_values: List[int]) -> int:

Inputs:
  - n (int): The number of distinct tokens available. 1 ≤ n ≤ 100.
  - x (int): The target total value to be achieved. 1 ≤ x ≤ 10^6.
  - token_values (List[int]): A list of integers representing the values of each token. 1 ≤ token_values[i] ≤ 10^6 for all i.

Returns:
  - int: The minimum number of tokens required to make the total x, or -1 if it is not possible to form the total with the given tokens.

Examples:
1. Input: n = 3, x = 11, token_values = [1, 5, 7]
  Output: 3
  Explanation: 
  The minimum number of tokens to make a total of 11 is three, using the combination 5 + 5 + 1.

2. Input: n = 3, x = 9, token_values = [2, 3, 5]
  Output: 3
  Explanation:
  The minimum number of tokens to make a total of 9 is three, using the combination 2 + 2 + 5.

Notes:
  - The problem can be approached using dynamic programming, where the state signifies the minimum number of tokens needed to achieve a certain total.
  - The algorithm iterates over the available token values, computing the minimum number of tokens required for increasing totals up to x.
  - If no combination of the given tokens can accumulate to the total x, the function should return -1.

Tags:
  - Dynamic Programming
  - Greedy Algorithm

"""

def min_tokens_for_total(n, x, token_values):
  pass